--- a/lib/assets.js
+++ b/lib/assets.js
@@ -15,16 +15,18 @@ TradeOfferManager.prototype._digestDescriptions = function(descriptions) {
 
 	if (descriptions && !(descriptions instanceof Array)) {
 		descriptions = Object.keys(descriptions).map(key => descriptions[key]);
-	}
 
-	(descriptions || []).forEach((item) => {
-		if (!item || !item.appid || !item.classid) {
-			return;
+		for (var i = 0; i < descriptions.length; i++) {
+			var item = descriptions[i];
+
+			if (!item || !item.appid || !item.classid) {
+				return;
+			}
+
+			cache.add(`${item.appid}_${item.classid}_${item.instanceid || '0'}`, item);
+			this._persistToDisk(`asset_${item.appid}_${item.classid}_${item.instanceid || '0'}.json`, JSON.stringify(item));
 		}
-
-		cache.add(`${item.appid}_${item.classid}_${item.instanceid || '0'}`, item);
-		this._persistToDisk(`asset_${item.appid}_${item.classid}_${item.instanceid || '0'}.json`, JSON.stringify(item));
-	});
+	}
 };
 
 TradeOfferManager.prototype._mapItemsToDescriptions = function(appid, contextid, items) {
@@ -83,7 +85,9 @@ TradeOfferManager.prototype._requestDescriptions = function(classes, callback) {
 		var appids = [];
 
 		// Split this out into appids
-		classes.forEach((item) => {
+		for(var i = 0; i < classes.length; i++) {
+			var item = classes[i];
+
 			// Don't add this if we already have it in the cache
 			if (this._assetCache.get(`${item.appid}_${item.classid}_${item.instanceid || '0'}`)) {
 				return;
@@ -101,7 +105,7 @@ TradeOfferManager.prototype._requestDescriptions = function(classes, callback) {
 			if (apps[index].indexOf(item.classid + '_' + (item.instanceid || '0')) == -1) {
 				apps[index].push(item.classid + '_' + (item.instanceid || '0'));
 			}
-		});
+		}
 
 		Async.map(apps, (app, cb) => {
 			var chunks = [];
@@ -119,11 +123,13 @@ TradeOfferManager.prototype._requestDescriptions = function(classes, callback) {
 					"class_count": chunk.length
 				};
 
-				chunk.forEach((item, index) => {
+				for (var i = 0; index < chunk.length; i++) {
+					var item = chunk[index];
 					var parts = item.split('_');
+
 					input['classid' + index] = parts[0];
 					input['instanceid' + index] = parts[1];
-				});
+				}
 
 				this.emit('debug', "Requesting classinfo for " + chunk.length + " items from app " + app.appid);
 				this._apiCall('GET', {
@@ -167,7 +173,11 @@ TradeOfferManager.prototype._requestDescriptions = function(classes, callback) {
 				return;
 			}
 
-			result.forEach(this._digestDescriptions.bind(this));
+			for (var i = 0; i < result.length; i++) {
+				var descriptions = result[i];
+				this._digestDescriptions(descriptions);
+			}
+
 			callback();
 		});
 	};