--- a/lib/classes/TradeOffer.js
+++ b/lib/classes/TradeOffer.js
@@ -170,11 +170,14 @@ TradeOffer.prototype.addMyItem = function(item) {
 TradeOffer.prototype.addMyItems = function(items) {
 	var added = 0;
 	var self = this;
-	items.forEach(function(item) {
+
+	for (var i = 0; i < items.length; i++) {
+		var item = items[i];
+
 		if (self.addMyItem(item)) {
 			added++;
 		}
-	});
+	}
 
 	return added;
 };
@@ -196,11 +199,14 @@ TradeOffer.prototype.removeMyItem = function(item) {
 
 TradeOffer.prototype.removeMyItems = function(items) {
 	var removed = 0;
-	items.forEach((item) => {
+
+	for (var i = 0; i < items.length; i++) {
+		var item = items[i];
+
 		if (this.removeMyItem(item)) {
 			removed++;
 		}
-	});
+	}
 
 	return removed;
 };
@@ -209,13 +215,16 @@ TradeOffer.prototype.addTheirItem = function(item) {
 	return addItem(item, this, this.itemsToReceive);
 };
 
-TradeOffer.prototype.addTheirItems = function(items) {
+TradeOffer.prototype.addTheirItems = function (items) {
 	var added = 0;
-	items.forEach((item) => {
+
+	for (var i = 0; i < items.length; i++) {
+		var item = items[i];
+
 		if (this.addTheirItem(item)) {
 			added++;
 		}
-	});
+	}
 
 	return added;
 };
@@ -237,11 +246,14 @@ TradeOffer.prototype.removeTheirItem = function(item) {
 
 TradeOffer.prototype.removeTheirItems = function(items) {
 	var removed = 0;
-	items.forEach((item) => {
+
+	for (var i = 0; i < items.length; i++) {
+		var item = items[i];
+
 		if (this.removeTheirItem(item)) {
 			removed++;
 		}
-	});
+	}
 
 	return removed;
 };