--- a/lib/helpers.js
+++ b/lib/helpers.js
@@ -71,20 +71,33 @@ exports.offerSuperMalformed = offerSuperMalformed;
 exports.offerMalformed = offerMalformed;
 exports.processItems = processItems;
 
-exports.checkNeededDescriptions = function(manager, offers, callback) {
+exports.checkNeededDescriptions = function (manager, offers, callback) {
 	if (!manager._language) {
 		callback(null);
 		return;
 	}
 
 	var items = [];
-	offers.forEach((offer) => {
-		(offer.items_to_give || []).concat(offer.items_to_receive || []).forEach((item) => {
+
+	var check = function (itemsToCheck) {
+		for (var i = 0; i < itemsToCheck.length; i++) {
+			var item = itemsToCheck[i];
+
 			if (!manager._hasDescription(item)) {
 				items.push(item);
 			}
-		});
-	});
+		}
+	}
+
+	for (var i = 0; i < offers.length; i++) {
+		var offer = offers[i];
+
+		var items_to_give = offer.items_to_give || [];
+		var items_to_receive = offer.items_to_receive || [];
+
+		check(items_to_give);
+		check(items_to_receive);
+	}
 
 	if (!items.length) {
 		callback(null);