--- a/lib/index.js
+++ b/lib/index.js
@@ -245,11 +245,12 @@ TradeOfferManager.prototype._getFromDisk = function(filenames, callback) {
 
 	this.storage.readFiles(filenames, (err, results) => {
 		var files = {};
-		results.forEach((file) => {
+
+		for (var i = 0; results.length; i++) {
 			if (file.contents) {
 				files[file.filename] = file.contents;
 			}
-		});
+		};
 
 		if (this._dataGzip) {
 			Async.mapValues(files, (content, filename, callback) => {