--- a/lib/polling.js
+++ b/lib/polling.js
@@ -115,7 +115,9 @@ TradeOfferManager.prototype.doPoll = function(doFullUpdate) {
 		var offers = this.pollData.sent || {};
 		var hasGlitchedOffer = false;
 
-		sent.forEach((offer) => {
+		for (var i = 0; i < sent.length; i++) {
+			var offer = sent[i];
+
 			if (!offers[offer.id]) {
 				// We sent this offer, but we have no record of it! Good job Steam
 				// Apparently offers can appear in the API before the send() call has returned, so we'll need to add a delay
@@ -148,7 +150,9 @@ TradeOfferManager.prototype.doPoll = function(doFullUpdate) {
 					timestamps[offer.id] = offer.created.getTime() / 1000;
 				} else {
 					hasGlitchedOffer = true;
-					var countWithoutName = !this._language ? 0 : offer.itemsToGive.concat(offer.itemsToReceive).filter(function(item) { return !item.name; }).length;
+					var countWithoutName = !this._language ? 0 : offer.itemsToGive.concat(offer.itemsToReceive).filter(function (item) {
+						return !item.name;
+					}).length;
 					this.emit('debug', "Not emitting sentOfferChanged for " + offer.id + " right now because it's glitched (" +
 						offer.itemsToGive.length + " to give, " + offer.itemsToReceive.length + " to receive, " + countWithoutName + " without name)");
 				}
@@ -194,7 +198,7 @@ TradeOfferManager.prototype.doPoll = function(doFullUpdate) {
 					});
 				}
 			}
-		});
+		}
 
 		if (this.cancelOfferCount) {
 			var sentActive = sent.filter(offer => offer.state == ETradeOfferState.Active);
@@ -223,7 +227,9 @@ TradeOfferManager.prototype.doPoll = function(doFullUpdate) {
 		this.pollData.sent = offers;
 		offers = this.pollData.received || {};
 
-		received.forEach((offer) => {
+		for (var i = 0; i < received.length; i++) {
+			var offer = received[i];
+
 			if (offer.isGlitched()) {
 				hasGlitchedOffer = true;
 				return;
@@ -246,7 +252,7 @@ TradeOfferManager.prototype.doPoll = function(doFullUpdate) {
 
 			offers[offer.id] = offer.state;
 			timestamps[offer.id] = offer.created.getTime() / 1000;
-		});
+		}
 
 		this.pollData.received = offers;
 		this.pollData.timestamps = timestamps;
@@ -254,12 +260,20 @@ TradeOfferManager.prototype.doPoll = function(doFullUpdate) {
 		// Find the latest update time
 		if (!hasGlitchedOffer) {
 			var latest = this.pollData.offersSince || 0;
-			sent.concat(received).forEach((offer) => {
-				var updated = Math.floor(offer.updated.getTime() / 1000);
-				if (updated > latest) {
-					latest = updated;
+
+			var compare = function (offers) {
+				for (var i = 0; i < offers.length; i++) {
+					var offer = offers[i];
+					var updated = Math.floor(offer.updated.getTime() / 1000);
+
+					if (updated > latest) {
+						latest = updated;
+					}
 				}
-			});
+			}
+
+			compare(sent);
+			compare(received);
 
 			this.pollData.offersSince = latest;
 		}