```
By simply rewriting .forEach, into loops, we would have a very nice increase in performance, as you can see in my test code
```

<details>
<summary>Test code</summary>

```javascript
"use strict";
console.log('initing performance test, node version:', process.version);
console.log('creating big array...');

var bigArray = new Array(10000000);
console.log(`array size: ${bigArray.length}`);

{
    console.time('forEach');
    bigArray.forEach(function (val) {
        //do something?
    });
    console.timeEnd('forEach');
}

{
    console.time('for');
    for (let i = 0; i < bigArray.length; i++) {
        //do something?
    }
    console.timeEnd('for');
}
```
</details>

<details>
<summary>Benchmarks</summary>
<br />

 <details>
<summary>14.18.2</summary>

```
initing performance test, node version: v14.18.2
creating big array...
array size: 10000000
forEach: 13.061ms
for: 5.085ms
```

</details>
 
<details>
<summary> v4.9.1</summary>

```
initing performance test, node version: v4.9.1
creating big array...
array size: 10000000
forEach: 610ms
for: 65ms
```

</details>

</details>