@DoctorMcKay 2022-02-21
```
initing performance test, node version: v14.17.3
creating big array...
array size: 1000000
forEach: 4.953ms
for: 6.267ms
```

The loop is actually slower for a more reasonable array size of 1 million.

```
initing performance test, node version: v14.17.3
creating big array...
array size: 100000
forEach: 0.651ms
for: 5.266ms
```

And **much** slower for an array of 100k.

Also, your naive replacement of forEach with for loops means that you left in return statements that should have been converted to continue statements.